Rename the market_gateway to market.
