+++ model/src/main/proto/spine_examples/shareaware/market/market.proto
@@ -24,12 +24,23 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * This package defines shares market gateway data types.
- */
-@CheckReturnValue
-@ParametersAreNonnullByDefault
-package io.spine.examples.shareaware.marketgateway;
+syntax = "proto3";
 
-import javax.annotation.CheckReturnValue;
-import javax.annotation.ParametersAreNonnullByDefault;
+package spine_examples.shareaware.market;
+
+import "spine/options.proto";
+
+option (type_url_prefix) = "type.shareaware.spine.io";
+option java_package = "io.spine.examples.shareaware.market";
+option java_outer_classname = "MarketProto";
+option java_multiple_files = true;
+
+import "spine_examples/shareaware/identifiers.proto";
+
+// The imitation of the shares market.
+message Market {
+    option (entity) = {kind: PROCESS_MANAGER};
+
+    // The ID of the shares market.
+    MarketId id = 1;
+}
