+++ model/src/main/proto/spine_examples/shareaware/market/rejections.proto
@@ -24,12 +24,24 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * Contains shares market gateway commands, and their interfaces.
- */
-@CheckReturnValue
-@ParametersAreNonnullByDefault
-package io.spine.examples.shareaware.marketgateway.command;
+syntax = "proto3";
 
-import javax.annotation.CheckReturnValue;
-import javax.annotation.ParametersAreNonnullByDefault;
+package spine_examples.shareaware.market;
+
+import "spine/options.proto";
+
+option (type_url_prefix) = "type.shareaware.spine.io";
+option java_package = "io.spine.examples.shareaware.market.rejection";
+option java_multiple_files = false;
+
+import "spine_examples/shareaware/identifiers.proto";
+
+// Shares cannot be obtained due to error in the market.
+message SharesCannotBeObtained {
+
+    // The ID of the purchase process that tried to obtain shares.
+    PurchaseId purchase_process = 1;
+
+    // Why the process failed.
+    string cause = 2;
+}
